requirements.txt
# Agent Negotiation System - Dependencies
# ========================================

# Core - Google ADK (THE SHELL)
google-adk>=1.0.0

# Schemas
pydantic>=2.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Orchestration (optional - can use ADK's workflow agents instead)
langgraph>=0.0.40

# Observability & Evaluation
langsmith>=0.1.0
langchain-core>=0.1.0

# LLM providers (optional - for LLM strategies beyond Gemini)
openai>=1.0.0

# Transport (optional - for WebSocket)
websockets>=12.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
